feat: add widget modules for window detection, photo display, and progress tracking
