+++ src/widget/Setting.py
@@ -1,0 +1,35 @@
+#重试次数
+RETRY = 3
+
+# 图片路径
+photo_dir = './assets/images/photo.png'
+
+#资源文件夹
+assets_dir = './assets'
+
+""" 
+图标路径
+"""
+
+# 上传
+window_detection_widget_icon= assets_dir + '/icon/window-detection.png'
+
+# 状态
+status_widget_icon = assets_dir + '/icon/check-status.png'
+
+# 照片
+photo_widget_icon = assets_dir + '/icon/photo.png'
+
+# 进度
+day_progress_widget_icon = assets_dir + '/icon/day-progress.png'
+
+
+"""
+布局文件
+"""
+
+# 基本
+
+base_widget_layout = assets_dir + '/ui/widget-base.ui'
+
+
